feat: author and place filter mobile versions
